Increase size of some atomic variables for better performance.
The Tile-Gx architecture has native support for 32-bit and 64-bit
atomic operations, but not 8-bit and 16-bit, which are emulated, so
changing some 16-bit and 8-bit atomic into ints improves performance.

Increasing the size of the atomic variables modified in this change
does not increase the total size of the structures in which they
reside because of existing padding requirements.

+++ src/flow.h
@@ -290,7 +290,7 @@ typedef struct Flow_
      *  On receiving a packet the counter is incremented while the flow
      *  bucked is locked, which is also the case on timeout pruning.
      */
-    SC_ATOMIC_DECLARE(unsigned short, use_cnt);
+    SC_ATOMIC_DECLARE(unsigned int, use_cnt);
 
     /** flow queue id, used with autofp */
     SC_ATOMIC_DECLARE(int, autofp_tmqh_flow_qid);
